refactor(helper-plugin): convert useAPIHandler to TS
